feat: first working version for sphere

+++ material.py
@@ -3,8 +3,8 @@ import numpy as np
 DIFFUSE = "diffuse"
 # Color values are not in uint8 so just make sure to clip everything to uint8
 # when rendering the final image
-COLOR_BLUE = np.array([10, 10, 230])
-COLOR_GRAY = np.array([135, 135, 135])
+COLOR_BLUE = np.array([10, 10, 230], dtype=float)
+COLOR_GRAY = np.array([135, 135, 135], dtype=float)
 
 
 class Material:
